+++ SIL.Windows.Forms.DblBundle/SelectProjectDlgBase.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.IO;
 using System.Windows.Forms;
 using L10NSharp;
@@ -8,7 +8,9 @@ namespace SIL.Windows.Forms.DblBundle
 {
 	public abstract class SelectProjectDlgBase : IDisposable
 	{
-		private readonly OpenFileDialog m_fileDialog;
+		private OpenFileDialog m_fileDialog;
+		private string m_defaultDir;
+		private bool m_allowProjectFiles;
 
 		protected abstract string DefaultBundleDirectory { get; set; }
 		protected abstract string ProjectFileExtension { get; }
@@ -18,20 +20,27 @@ namespace SIL.Windows.Forms.DblBundle
 		protected SelectProjectDlgBase(bool allowProjectFiles = true, string defaultFile = null)
 		{
 			FileName = File.Exists(defaultFile) ? Path.GetFileName(defaultFile) : null;
-			var defaultDir = (defaultFile != null ? Path.GetDirectoryName(defaultFile) : DefaultBundleDirectory);
-			if (string.IsNullOrEmpty(defaultDir) || !Directory.Exists(defaultDir))
+			m_defaultDir = (defaultFile != null ? Path.GetDirectoryName(defaultFile) : DefaultBundleDirectory);
+
+			//avoid default value (false)
+			m_allowProjectFiles = allowProjectFiles;
+		}
+
+		protected virtual OpenFileDialog CreateFileDialog()
+		{
+			if (string.IsNullOrEmpty(m_defaultDir) || !Directory.Exists(m_defaultDir))
 			{
-				defaultDir = Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments);
+				m_defaultDir = Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments);
 			}
 			string projectFiles = "";
-			if (allowProjectFiles)
+			if (m_allowProjectFiles)
 				projectFiles = string.Format("{0} ({1})|{1}|",
 					string.Format(LocalizationManager.GetString("DialogBoxes.SelectProjectDlg.ProjectFilesLabel", "{0} Project Files", "{0} is the product name"), ProductName),
 					"*" + ProjectFileExtension);
 			m_fileDialog = new OpenFileDialog
 			{
 				Title = Title,
-				InitialDirectory = defaultDir,
+				InitialDirectory = m_defaultDir,
 				FileName = FileName,
 				Filter = string.Format("{0} ({1})|{1}|{2}{3} ({4})|{4}",
 					LocalizationManager.GetString("DialogBoxes.SelectProjectDlg.ResourceBundleFileTypeLabel", "Text Resource Bundle files"),
@@ -41,10 +50,13 @@ namespace SIL.Windows.Forms.DblBundle
 					"*.*"),
 				DefaultExt = DblBundleFileUtils.kDblBundleExtension
 			};
+
+			return m_fileDialog;
 		}
 
 		public DialogResult ShowDialog()
 		{
+			m_fileDialog = CreateFileDialog();
 			var result = m_fileDialog.ShowDialog();
 			if (result == DialogResult.OK)
 			{
@@ -60,7 +72,17 @@ namespace SIL.Windows.Forms.DblBundle
 
 		public void Dispose()
 		{
-			m_fileDialog.Dispose();
+			Dispose(true);
+			GC.SuppressFinalize(this);
+		}
+
+		protected virtual void Dispose(bool disposing)
+		{
+			if (disposing)
+			{
+				m_fileDialog.Dispose();
+				m_fileDialog = null; 
+			}
 		}
 	}
 }